went/going through storyController and sought to update functions as applicable :) .  I have yet to fix all the errors and code breaks :)

+++ HNT_Back/controllers/storyController.js
@@ -1,5 +1,6 @@
 const Story = require('../model/Story');
 const User = require('../model/User');
+const Profile = require('../model/Profile')
 const StoryRating = require('../model/StoryRating');
 const Tag = require('../model/Tag');
 const Playlist = require('../model/Playlist');
@@ -31,7 +32,6 @@ const getStory = async (req, res) => {
     res.json(story);
   }
 }
-
 const getStories = async (req, res) => {
   // console.log('getStories backend!')
   
@@ -46,31 +46,43 @@ const getStories = async (req, res) => {
   res.json(stories);
 }
 
+
+
 //* Create or update a single story
 const saveStory = async (req, res) => {
+  const user = await User.findOne({ username: req.user});
+  const user_id = user._id;
+  const body = req.body;
   const story_id = req.body.story_id;
-  const new_story_info = {
-      id: req.body.story_id,
-      account_id: req.body.account_id,
-      audio_url: req.body.audio_url,
-      title: req.body.title,
-      tag_names: req.body.tag_names, 
-      description: req.body.description,
-      duration: req.body.duration,
-      private: req.body.private, 
-    };
+  const { violenceRating, sexRating, languageRating, generalRating, 
+    title, description, isPrivate, tags: rawTags, selectedFile, audioLink } = body;
+  const duration = 1000; //todo later we'll get this value from the file length or whatnot.
+  
+  let jsonnedTags = JSON.parse(rawTags);
+  const tags = jsonnedTags.map((tag)=> properlyUppercased(tag));
 
-  // Find the document
-  const options = { upsert: true, new: true, setDefaultsOnInsert: true };
-
-  const story = await Story.findOneAndUpdate({id: story_id}, new_story_info, options, function(error, result) {
-      if (error) return;
-      // do something with the document
-      res.status(200);
+  tags.map(async (tag) => { //@ b) saving tags to tags collection (?)
+      await Tag.findOneAndUpdate({ name: tag }, { name: tag }, { upsert: true});
   });
 
-}
+  const new_story_info = { //@ c) create the story on the database
+    user_id, title, description, isPrivate, tags, audioLink, duration, 
+  };
 
+  const options = { upsert: true, new: true, setDefaultsOnInsert: true };
+  // const story = await Story.findOneAndUpdate({ _id: story_id}, new_story_info, options, function(error, result) {
+  //     if (error) return;
+  //     // do something with the document
+  //     res.status(200);
+  // });
+  const story = await Story.findOneAndUpdate({ _id: story_id}, new_story_info, options);
+
+  const rating_info = { user_id, story_id, violenceRating, sexRating, languageRating, generalRating }
+  //?  They're only updating the rating they themselves are responsible for, right?
+  let ratings = await StoryRating.findOneAndUpdate({ _id: story_id}, rating_info, options);
+  
+  res.json({ story, ratings });
+}
 //* Create a new story
 const create = async (req, res) => {
   const user = await User.findOne({ username: req.user});
@@ -86,16 +98,16 @@ const create = async (req, res) => {
   let jsonnedTags = JSON.parse(rawTags);
   const tags = jsonnedTags.map((tag)=> properlyUppercased(tag));
 
-  console.log('tags are: ', tags); //okay
+  // console.log('tags are: ', tags);
 
   tags.map(async (tag) => { //@ b) saving tags to tags collection (?)
       await Tag.findOneAndUpdate({ name: tag }, { name: tag }, { upsert: true});
   });
 
   const new_story_info = { //@ c) create the story on the database
-    user_id: user_id, title, description, isPrivate, tags, audioLink, duration, 
-  }; //todo currently using user for user_id, as user_id is undefined!
-  //* the last part should create the createdAt and updatedAt sections, so to speak :)
+    user_id, title, description, isPrivate, tags, audioLink, duration, 
+  };
+  
   let story = await Story.create(new_story_info); 
   const story_id = story._id;
   console.log("story._id is: ", story._id); 
@@ -106,7 +118,7 @@ const create = async (req, res) => {
   console.log(story.createdAt, ", " + story.updatedAt);
   
   const rating_info = {  //@ d) create the ratings object in the database
-    user_id: user_id, story_id, violenceRating, sexRating, languageRating, generalRating 
+    user_id, story_id, violenceRating, sexRating, languageRating, generalRating 
   } //* NOTE: enjoymentRating not put in by creator :D
   let ratings = await StoryRating.create(rating_info);
   
@@ -118,12 +130,16 @@ const create = async (req, res) => {
 const popular = async (req, res) => {
   let stories = await Story.find({}) // sort by popularity_rating
   res.json(stories);
-};
+  //? calculate by popularity somewhere (?)
+}; 
 
 //todo Get most popular stories for a tag
 const popularByTag = async (req, res) => {
-  const tag_name = req.params.tag_name;
-  let stories = await Story.find({tag_name: tag_name}, ) // sort by popularity_rating
+  const { tags: rawTags } = req.params;
+  let jsonnedTags = JSON.parse(rawTags);
+  const tags = jsonnedTags.map((tag)=> properlyUppercased(tag));
+  //? calculate by popularity somewhere (?)
+  let stories = await Story.find({ tags }) // sort by popularity_rating
   res.json(stories);
 };
 
@@ -132,13 +148,17 @@ const search = async (req, res) => {
   const search_string = req.params.search_string;
 
   //? Get tags that match
-  let tag_stories = await Story.find({tag_names: search_string});
-
+  const search_tags = search_string.split(",").trim();
+  console.log('search_tags are: ', search_tags);
+  let tag_stories = await Story.find({ tags: { $in: search_tags } });
+  console.log('tag_stories are: ', tag_stories);
+  // .findOne( { topics : {$in: ['voyage', 'nautical']} }, { title: 1 } )
   //? Get creator accounts that match name
-  let creator_accounts = await Creator.find({name: search_string});
+  //? there is no Creator model.  Did you mean User or Profile?
+  let creator_accounts = await Profile.find({ name: search_string }); //Creator
 
   //? get titles that match
-  let title_stories = await Story.find({name: search_string});
+  let title_stories = await Story.find({ title: search_string });
 
   let search_results = {
     tag_matched_stories: tag_stories,
@@ -148,32 +168,22 @@ const search = async (req, res) => {
   res.json(search_results);
 };
 
+
+
 //* Get stories by playlist
 const storiesForPlaylist = async (req, res) => {
   const playlist_id = req.params.playlist_id;
   let playlist_stories = [];
 
-  let playlist = await Playlist.findOne({_id: playlist_id});
+  let playlist = await Playlist.findOne({ _id: playlist_id }); //* returns single playlist
   let playlist_story_ids = playlist.story_ids;
 
-
   playlist_story_ids.forEach(async (story_id) => {
-    let story = await Story.findOne({_id: story_id});
+    let story = await Story.findOne({ _id: story_id });
     playlist_stories.push(story);
-    res.json(playlist_stories);
   });
-  //? Lonnie's Incoming 6/17
-  // playlist_story_ids.forEach((story_id) => {
-  //   // let story = await Story.findOne({_id: story_id});
-  //   Story.findOne({_id: story_id})
-  //   .then((story) => {
-  //     playlist_stories.push(story);
-  //   })
-  // })
-  // .then(() => {
-  //   res.json(playlist_stories);
-  // })
-  
+
+  res.json(playlist_stories);
 };
 
 //0) playlist_id
@@ -181,11 +191,11 @@ const storiesForPlaylist = async (req, res) => {
 
 //* Get stories by creator
 const storiesByCreator = async (req, res) => {
-  let user = await User.findOne({ username: req.user });
+  const user = await User.findOne({ username: req.user });
   const user_id = user._id;
   console.log('storiesByCreator 184!')
-  
-  let stories = await Story.find({user_id: user_id});
+
+  let stories = await Story.find({ user_id: user_id });
   console.log('user: ', user, ', stories: ', stories)
   res.json(stories);
 };
@@ -194,32 +204,45 @@ const storiesByCreator = async (req, res) => {
 const show = async (req, res) => {
   const story_id = req.params.story_id;
   let story = await Story.findOne({_id: story_id});
+
   res.json(story);
 };
 
 //* Update an existing story
 const update = async (req, res) => {
+  const user = await User.findOne({ username: req.user });
+  const user_id = user._id;
   const story_id = req.params.story_id;
-  const story_data = {
-      id: req.body.story_id,
-      account_id: req.body.account_id,
-      audio_url: req.body.audio_url,
-      title: req.body.title,
-      tag_names: req.body.tag_names, 
-      description: req.body.description,
-      duration: req.body.duration,
-      private: req.body.private, 
-    };
 
-  const story = await Story.findOneAndUpdate({_id: story_id}, story_data, {upsert: true});
+  const body = req.body;
+  const { title, description, isPrivate, tags: rawTags, selectedFile, audioLink } = body;
+  const duration = 1000; //todo later we'll get this value from the file length or whatnot.
+  
+  let jsonnedTags = JSON.parse(rawTags);
+  const tags = jsonnedTags.map((tag)=> properlyUppercased(tag));
+
+  tags.map(async (tag) => { //@ b) saving tags to tags collection (?)
+    await Tag.findOneAndUpdate({ name: tag }, { name: tag }, { upsert: true});
+  });
+
+  const story_data = { //@ c) create the story on the database
+    user_id, title, description, isPrivate, tags, audioLink, duration, 
+  };
+
+  const story = await Story.findOneAndUpdate({ _id: story_id, user_id }, story_data, {upsert: true});
   res.json(story);
 };
 
 //* Delete a story
 const remove = async (req, res) => {
+  const user = await User.findOne({ username: req.user});
+  const user_id = user._id;
+
   const story_id = req.params.story_id;
 
-  let story = await Story.findOneAndDelete({_id: story_id});
+  //? Do we want to make sure it's the right user who is deleting the story?  If not, 
+  //... remove the user_id requirement!
+  let story = await Story.findOneAndDelete({ _id: story_id, user_id: user_id });
   res.json(story);
 };
 
@@ -242,3 +265,19 @@ module.exports = {
 
   // Total stars from all the ratings a story received?
     // Should we consider time, like the past 30 days?
+
+
+
+
+    
+  //? Lonnie's Incoming 6/17
+  // playlist_story_ids.forEach((story_id) => {
+  //   // let story = await Story.findOne({_id: story_id});
+  //   Story.findOne({_id: story_id})
+  //   .then((story) => {
+  //     playlist_stories.push(story);
+  //   })
+  // })
+  // .then(() => {
+  //   res.json(playlist_stories);
+  // })
